requirements.txt
appdirs==1.4.3
Django==1.10
django-betterforms==1.1.4
django-cors-middleware==1.3.1
django-timezone-field==1.2
djangorestframework==3.6.2
jsonfield==2.0.1
mysqlclient==1.3.10
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
urllib3==1.20
